third_party/blink/web_tests/fast/forms/label/contenteditable-label-focus-tab-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

Focus should be on the label
PASS document.activeElement is label
Focus should be on the button after pressing tab
PASS document.activeElement is button
Focus should be on the label after pressing shift+tab
PASS document.activeElement is label
I'm editable. Just a button
